Refactor CoreAudioTapEngine to improve buffer processing and error handling. Consolidate audio format conversion methods and enhance error management with new cases in CoreAudioTapEngineError. Remove redundant code and improve clarity in audio processing logic.
