+++ Livcap/CoreAudioTapEngine/CoreAudioTapEngineError.swift
@@ -16,9 +16,12 @@ enum CoreAudioTapEngineError: Error, LocalizedError {
     case aggregateCreationFailed(OSStatus)
     case formatCreationFailed
     case invalidTapStreamDescription
+    case invalidStreamDescription
     case ioProcCreationFailed(OSStatus)
     case deviceStartFailed(OSStatus)
     case tapNotInstalled
+    case unsupportedMacOSVersion
+    case processNotFound(String)
 
     var errorDescription: String? {
         switch self {
@@ -32,12 +35,18 @@ enum CoreAudioTapEngineError: Error, LocalizedError {
             return "Unable to create AVAudioFormat from tap stream description."
         case .invalidTapStreamDescription:
             return "Invalid or missing stream description from the audio tap."
+        case .invalidStreamDescription:
+            return "Invalid audio stream description."
         case .ioProcCreationFailed(let status):
             return "Failed to create I/O proc for audio processing. Error code: \(status)"
         case .deviceStartFailed(let status):
             return "Failed to start audio device for capture. Error code: \(status)"
         case .tapNotInstalled:
             return "Tap must be installed before starting the engine."
+        case .unsupportedMacOSVersion:
+            return "System audio capture requires macOS 14.4 or later."
+        case .processNotFound(let processName):
+            return "Process not found: \(processName). Please ensure it is running."
         }
     }
 }
