+++ Livcap/Service/SystemAudioManager.swift
@@ -732,7 +732,7 @@ class SystemAudioManager: ObservableObject, SystemAudioProtocol {
         // NEW: Find Chrome Helper ObjectIDs with precise bundle targeting
         let chromeHelperObjectIDs = findChromeHelperObjectIDs()
         guard !chromeHelperObjectIDs.isEmpty else {
-            throw SystemAudioError.processNotFound("Google Chrome Helper (com.google.Chrome.helper)")
+            throw CoreAudioTapEngineError.processNotFound("Google Chrome Helper (com.google.Chrome.helper)")
         }
 
         logger.info("🚀 TARGETING CHROME HELPER OBJECT IDs: \(chromeHelperObjectIDs)")
@@ -749,7 +749,7 @@ class SystemAudioManager: ObservableObject, SystemAudioProtocol {
         var tapID: AUAudioObjectID = .unknown
         let err = AudioHardwareCreateProcessTap(tapDescription, &tapID)
         guard err == noErr else {
-            throw SystemAudioError.tapCreationFailed(err)
+            throw CoreAudioTapEngineError.tapCreationFailed(err)
         }
 
         self.processTapID = tapID
@@ -782,7 +782,7 @@ class SystemAudioManager: ObservableObject, SystemAudioProtocol {
         var aggregateID = AudioObjectID.unknown
         let aggregateErr = AudioHardwareCreateAggregateDevice(description as CFDictionary, &aggregateID)
         guard aggregateErr == noErr else {
-            throw SystemAudioError.aggregateDeviceCreationFailed(aggregateErr)
+            throw CoreAudioTapEngineError.aggregateCreationFailed(aggregateErr)
         }
 
         self.aggregateDeviceID = aggregateID
@@ -794,11 +794,11 @@ class SystemAudioManager: ObservableObject, SystemAudioProtocol {
     
     private func setupAudioProcessing() throws {
         guard var streamDescription = tapStreamDescription else {
-            throw SystemAudioError.invalidStreamDescription
+            throw CoreAudioTapEngineError.invalidStreamDescription
         }
         
         guard let inputFormat = AVAudioFormat(streamDescription: &streamDescription) else {
-            throw SystemAudioError.formatCreationFailed
+            throw CoreAudioTapEngineError.formatCreationFailed
         }
         
         logger.info("Input format: \(inputFormat)")
@@ -824,7 +824,7 @@ class SystemAudioManager: ObservableObject, SystemAudioProtocol {
             }
             
             // Convert to target format if needed
-            let processedBuffer = self.convertBufferFormat(inputBuffer, to: targetFormatCapture)
+            let processedBuffer = convertBufferFormat(inputBuffer, to: targetFormatCapture)
             
             // Extract float samples and convert stereo to mono
             let frameCount = Int(processedBuffer.frameLength)
@@ -899,86 +899,19 @@ class SystemAudioManager: ObservableObject, SystemAudioProtocol {
         // Install I/O proc
         var err = AudioDeviceCreateIOProcIDWithBlock(&deviceProcID, aggregateDeviceID, queue, ioBlock)
         guard err == noErr else {
-            throw SystemAudioError.ioProcCreationFailed(err)
+            throw CoreAudioTapEngineError.ioProcCreationFailed(err)
         }
         
         // Start audio device
         err = AudioDeviceStart(aggregateDeviceID, deviceProcID)
         guard err == noErr else {
-            throw SystemAudioError.deviceStartFailed(err)
+            throw CoreAudioTapEngineError.deviceStartFailed(err)
         }
         
         logger.info("Audio processing started successfully")
         print("✅ SYSTEM AUDIO TAP IS ACTIVE - Should start receiving buffer data now")
     }
     
-    // MARK: - Audio Format Conversion
-    
-    private func convertBufferFormat(_ buffer: AVAudioPCMBuffer, to format: AVAudioFormat) -> AVAudioPCMBuffer {
-        // If formats match exactly, return original buffer
-        if buffer.format.sampleRate == format.sampleRate && 
-           buffer.format.channelCount == format.channelCount {
-            return buffer
-        }
-        
-
-        logger.info("💻  Coverting From: \(buffer.format.channelCount) ch, \(buffer.format.sampleRate) Hz ,  To: \(format.channelCount) ch, \(format.sampleRate) Hz")
-        
-        // For stereo-to-mono conversion, we'll handle it manually in the extraction phase
-        // Here we only handle sample rate conversion if needed, keeping original channel count
-        let intermediateFormat: AVAudioFormat
-        if buffer.format.sampleRate != format.sampleRate {
-            // Create intermediate format with same channels but target sample rate
-            guard let tempFormat = AVAudioFormat(
-                commonFormat: .pcmFormatFloat32,
-                sampleRate: format.sampleRate,
-                channels: buffer.format.channelCount,
-                interleaved: false
-            ) else {
-                logger.warning("Failed to create intermediate format, using original buffer")
-                return buffer
-            }
-            intermediateFormat = tempFormat
-        } else {
-            // No sample rate conversion needed, return original for channel conversion in extraction
-            return buffer
-        }
-        
-        // Create converter for sample rate conversion only
-        guard let converter = AVAudioConverter(from: buffer.format, to: intermediateFormat) else {
-            logger.warning("Failed to create sample rate converter, using original buffer")
-            return buffer
-        }
-        
-        // Calculate output frame capacity for sample rate conversion
-        let outputFrameCapacity = AVAudioFrameCount(
-            (Double(buffer.frameLength) / buffer.format.sampleRate) * intermediateFormat.sampleRate
-        )
-        
-        guard let outputBuffer = AVAudioPCMBuffer(
-            pcmFormat: intermediateFormat,
-            frameCapacity: outputFrameCapacity
-        ) else {
-            logger.warning("Failed to create output buffer, using original buffer")
-            return buffer
-        }
-        
-        outputBuffer.frameLength = outputFrameCapacity
-        
-        // Perform sample rate conversion
-        var error: NSError?
-        converter.convert(to: outputBuffer, error: &error) { _, outStatus in
-            outStatus.pointee = .haveData
-            return buffer
-        }
-        
-        if let error = error {
-            logger.warning("Sample rate conversion failed: \(error.localizedDescription), using original buffer")
-            return buffer
-        }
-        
-        return outputBuffer
-    }
     
     // MARK: - Cleanup
     
@@ -1125,208 +1058,10 @@ class SystemAudioManager: ObservableObject, SystemAudioProtocol {
     }
 }
 
-// MARK: - Error Types
 
-enum SystemAudioError: Error, LocalizedError {
-    case tapCreationFailed(OSStatus)
-    case aggregateDeviceCreationFailed(OSStatus)
-    case invalidStreamDescription
-    case formatCreationFailed
-    case ioProcCreationFailed(OSStatus)
-    case deviceStartFailed(OSStatus)
-    case unsupportedMacOSVersion
-    case processNotFound(String)
-    
-    var errorDescription: String? {
-        switch self {
-        case .tapCreationFailed(let status):
-            return "Failed to create audio tap: \(status)"
-        case .aggregateDeviceCreationFailed(let status):
-            return "Failed to create aggregate device: \(status)"
-        case .invalidStreamDescription:
-            return "Invalid audio stream description"
-        case .formatCreationFailed:
-            return "Failed to create audio format"
-        case .ioProcCreationFailed(let status):
-            return "Failed to create I/O proc: \(status)"
-        case .deviceStartFailed(let status):
-            return "Failed to start audio device: \(status)"
-        case .unsupportedMacOSVersion:
-            return "System audio capture requires macOS 14.4 or later"
-        case .processNotFound(let processName):
-            return "Process not found: \(processName). Please ensure it is running."
-        }
-    }
-}
 
-// MARK: - AudioObjectID Extensions (from AudioCap example)
+// MARK: - Permission Status Enum
 
-extension AudioObjectID {
-    /// Convenience for `kAudioObjectSystemObject`.
-    static let system = AudioObjectID(kAudioObjectSystemObject)
-    /// Convenience for `kAudioObjectUnknown`.
-    static let unknown = kAudioObjectUnknown
-
-    /// `true` if this object has the value of `kAudioObjectUnknown`.
-    var isUnknown: Bool { self == .unknown }
-
-    /// `false` if this object has the value of `kAudioObjectUnknown`.
-    var isValid: Bool { !isUnknown }
-}
-
-extension AudioObjectID {
-    /// Reads the process list from the system audio object
-    func readProcessList() throws -> [AudioObjectID] {
-        var address = AudioObjectPropertyAddress(
-            mSelector: kAudioHardwarePropertyProcessObjectList,
-            mScope: kAudioObjectPropertyScopeGlobal,
-            mElement: kAudioObjectPropertyElementMain
-        )
-        
-        var dataSize: UInt32 = 0
-        var err = AudioObjectGetPropertyDataSize(self, &address, 0, nil, &dataSize)
-        guard err == noErr else { throw "Failed to get process list size: \(err)" }
-        
-        let count = Int(dataSize) / MemoryLayout<AudioObjectID>.size
-        var value = [AudioObjectID](repeating: .unknown, count: count)
-        
-        err = AudioObjectGetPropertyData(self, &address, 0, nil, &dataSize, &value)
-        guard err == noErr else { throw "Failed to read process list: \(err)" }
-        
-        return value
-    }
-    
-    /// Reads the PID for a process object
-    func readPID() throws -> pid_t {
-        var address = AudioObjectPropertyAddress(
-            mSelector: kAudioProcessPropertyPID,
-            mScope: kAudioObjectPropertyScopeGlobal,
-            mElement: kAudioObjectPropertyElementMain
-        )
-        
-        var dataSize: UInt32 = UInt32(MemoryLayout<pid_t>.size)
-        var pid: pid_t = -1
-        
-        let err = AudioObjectGetPropertyData(self, &address, 0, nil, &dataSize, &pid)
-        guard err == noErr else { throw "Failed to read PID: \(err)" }
-        
-        return pid
-    }
-    
-    /// Reads the bundle ID for a process object
-    func readProcessBundleID() -> String? {
-        var address = AudioObjectPropertyAddress(
-            mSelector: kAudioProcessPropertyBundleID,
-            mScope: kAudioObjectPropertyScopeGlobal,
-            mElement: kAudioObjectPropertyElementMain
-        )
-        
-        var dataSize: UInt32 = 0
-        let err = AudioObjectGetPropertyDataSize(self, &address, 0, nil, &dataSize)
-        guard err == noErr, dataSize > 0 else { return nil }
-        
-        var cfString: CFString = "" as CFString
-        let err2 = AudioObjectGetPropertyData(self, &address, 0, nil, &dataSize, &cfString)
-        guard err2 == noErr else { return nil }
-        
-        let result = cfString as String
-        return result.isEmpty ? nil : result
-    }
-    
-    /// Reads whether the process is currently running
-    func readProcessIsRunning() -> Bool {
-        var address = AudioObjectPropertyAddress(
-            mSelector: kAudioProcessPropertyIsRunning,
-            mScope: kAudioObjectPropertyScopeGlobal,
-            mElement: kAudioObjectPropertyElementMain
-        )
-        
-        var dataSize: UInt32 = UInt32(MemoryLayout<UInt32>.size)
-        var value: UInt32 = 0
-        
-        let err = AudioObjectGetPropertyData(self, &address, 0, nil, &dataSize, &value)
-        return err == noErr && value == 1
-    }
-    
-    /// Reads audio status for a process (input/output)
-    func readProcessAudioStatus(_ property: AudioObjectPropertySelector) -> Bool {
-        var address = AudioObjectPropertyAddress(
-            mSelector: property,
-            mScope: kAudioObjectPropertyScopeGlobal,
-            mElement: kAudioObjectPropertyElementMain
-        )
-        
-        var dataSize: UInt32 = UInt32(MemoryLayout<UInt32>.size)
-        var value: UInt32 = 0
-        
-        let err = AudioObjectGetPropertyData(self, &address, 0, nil, &dataSize, &value)
-        return err == noErr && value == 1
-    }
-    
-    /// Reads the value for `kAudioHardwarePropertyDefaultSystemOutputDevice`.
-    static func readDefaultSystemOutputDevice() throws -> AudioDeviceID {
-        try AudioDeviceID.system.readDefaultSystemOutputDevice()
-    }
-    
-    /// Reads the value for `kAudioHardwarePropertyDefaultSystemOutputDevice`, should only be called on the system object.
-    func readDefaultSystemOutputDevice() throws -> AudioDeviceID {
-        try requireSystemObject()
-        return try read(kAudioHardwarePropertyDefaultSystemOutputDevice, defaultValue: AudioDeviceID.unknown)
-    }
-    
-    /// Reads the value for `kAudioDevicePropertyDeviceUID` for the device represented by this audio object ID.
-    func readDeviceUID() throws -> String {
-        try readString(kAudioDevicePropertyDeviceUID)
-    }
-    
-    /// Reads the value for `kAudioTapPropertyFormat` for the device represented by this audio object ID.
-    func readAudioTapStreamBasicDescription() throws -> AudioStreamBasicDescription {
-        try read(kAudioTapPropertyFormat, defaultValue: AudioStreamBasicDescription())
-    }
-    
-    private func requireSystemObject() throws {
-        if self != .system { throw "Only supported for the system object." }
-    }
-    
-    // Generic property access methods
-    func read<T>(_ selector: AudioObjectPropertySelector,
-                scope: AudioObjectPropertyScope = kAudioObjectPropertyScopeGlobal,
-                element: AudioObjectPropertyElement = kAudioObjectPropertyElementMain,
-                defaultValue: T) throws -> T {
-        try read(AudioObjectPropertyAddress(mSelector: selector, mScope: scope, mElement: element), defaultValue: defaultValue)
-    }
-    
-    func readString(_ selector: AudioObjectPropertySelector,
-                   scope: AudioObjectPropertyScope = kAudioObjectPropertyScopeGlobal,
-                   element: AudioObjectPropertyElement = kAudioObjectPropertyElementMain) throws -> String {
-        try read(AudioObjectPropertyAddress(mSelector: selector, mScope: scope, mElement: element), defaultValue: "" as CFString) as String
-    }
-    
-    private func read<T>(_ address: AudioObjectPropertyAddress, defaultValue: T) throws -> T {
-        var inAddress = address
-        var dataSize: UInt32 = 0
-        
-        var err = AudioObjectGetPropertyDataSize(self, &inAddress, 0, nil, &dataSize)
-        guard err == noErr else {
-            throw "Error reading data size for \(address): \(err)"
-        }
-        
-        var value: T = defaultValue
-        err = withUnsafeMutablePointer(to: &value) { ptr in
-            AudioObjectGetPropertyData(self, &inAddress, 0, nil, &dataSize, ptr)
-        }
-        
-        guard err == noErr else {
-            throw "Error reading data for \(address): \(err)"
-        }
-        
-        return value
-    }
-}
-
-extension String: @retroactive LocalizedError {
-    public var errorDescription: String? { self }
-}
 
 // MARK: - Permission Status Enum
 
@@ -1363,37 +1098,4 @@ enum PermissionStatus: String, CaseIterable {
     }
 }
 
-extension AudioDeviceID {
-    func getDeviceName() -> String {
-        var address = AudioObjectPropertyAddress(mSelector: kAudioObjectPropertyElementName, mScope: kAudioObjectPropertyScopeGlobal, mElement: kAudioObjectPropertyElementMain)
-        var dataSize: UInt32 = 0
-        var name: CFString = "" as CFString
-        
-        let err1 = AudioObjectGetPropertyDataSize(self, &address, 0, nil, &dataSize)
-        if err1 == noErr {
-            let err2 = AudioObjectGetPropertyData(self, &address, 0, nil, &dataSize, &name)
-            if err2 == noErr {
-                return name as String
-            }
-        }
-        
-        return "Unknown Device"
-    }
-}
-
-// Helper to get the default output device ID
-private func getDefaultOutputDeviceID() throws -> AudioDeviceID {
-    var address = AudioObjectPropertyAddress(mSelector: kAudioHardwarePropertyDefaultOutputDevice, mScope: kAudioObjectPropertyScopeGlobal, mElement: kAudioObjectPropertyElementMain)
-    var dataSize: UInt32 = 0
-    var deviceID: AudioDeviceID = 0
-    
-    let err = AudioObjectGetPropertyDataSize(AudioObjectID.system, &address, 0, nil, &dataSize)
-    if err == noErr {
-        let err2 = AudioObjectGetPropertyData(AudioObjectID.system, &address, 0, nil, &dataSize, &deviceID)
-        if err2 == noErr {
-            return deviceID
-        }
-    }
-    
-    throw SystemAudioError.processNotFound("Failed to get default output device ID")
-} 
+
